trp1-ai-artist/pyproject.toml
[project]
name = "ai-content"
version = "0.1.0"
description = "Multi-provider AI content generation framework for music, video, and images."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "10Academy Training", email = "training@10academy.org" }]
keywords = ["ai", "content-generation", "music", "video", "multimodal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    # Core
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    
    # Google APIs
    "google-genai>=1.51.0",
    "google-api-python-client>=2.120.0",
    "google-auth-oauthlib>=1.2.0",
    
    # Auth & Security
    "pyjwt>=2.8.0",
    
    # Media Processing
    "pillow>=10.2.0",
    
    # Legacy (for backwards compatibility)
    "internetarchive>=4.0.0",
    "lumaai>=0.6.0",
    "moviepy>=2.0.0.dev2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
ai-content = "ai_content.cli.main:cli"

[project.urls]
Repository = "https://github.com/10academy/ai-content"
Documentation = "https://github.com/10academy/ai-content#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_content"]

[tool.uv]
managed = true

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["ai_content"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "google.*",
    "lumaai.*",
    "internetarchive.*",
    "moviepy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
